Replace implementation of MediaDao to delegate to repositories
