--- media/src/test/java/fr/nihilus/music/media/os/SimulatedFileSystem.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright 2021 Thibault Seisel
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package fr.nihilus.music.media.os
-
-import fr.nihilus.music.core.os.FileSystem
-
-/**
- * A simulated file system used for testing purposes.
- * No actual files are stored.
- */
-internal class SimulatedFileSystem(
-    vararg filenames: String
-) : FileSystem {
-    private val storedFiles = filenames.toMutableSet()
-
-    override fun deleteFile(filepath: String): Boolean = storedFiles.remove(filepath)
-
-    fun fileExists(filepath: String): Boolean = storedFiles.contains(filepath)
-}
